requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.7
django-allauth==0.34.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
psycopg2==2.7
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
